Add experimental try/catch that only supports ExceptionInfo
Add trivial support for

  (try
    ...
  (catch ExceptionInfo ex
    ...)

See the README and DESIGN changes for additional information.

Because we don't want Guile's built-in catch to interfere with the
catch literal in the try define-syntax, make (lokke exception) a pure:
module and then explicitly import all the normal bindings except
catch.  There might be a better way to do this in the define-module,
but we can always fix that later.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>