Don't print anything for *unspecified*
In Guile *unspecified* is what's actually returned from expressions
like (if #f ""), and it's not a false value.  Change ./lokke to treat
it like nil when deciding whether or not to print "-e" expression
values, and arrange for its printed representation to be
nothing (i.e. the empty string), rather than "#object[<unknown>
0x804]".

Signed-off-by: Rob Browning <rlb@defaultvalue.org>