First cut at print support
This commit refactors print support using a polymorphic `print-on`
function. All print functions such as print, println, print-str bottom
out in the new `print-on` function. There will eventually be a `pr-on`
equivalent, but now `pr-on` just calls `print-on` to match previous
behavior and until eval round tripping can be verified/implemented.

Signed-off-by: Ryan Senior <senior.ryan@gmail.com>