Drop srfi-88 (postfix) keywords
Not sure why I imagined otherwise, but (reader-set! keywords ...)  is
not only global, but doesn't "reset" at the end of a module/file,
which means that it'll affect any subsequent (use-modules ...), and
there's no reason to expect that any given module won't break,
e.g. (web client).

That particular module breaks with a 'postfix setting because it has a
during compilation, so you won't see the problem if the module has
already been compiled.  Note also that loading (srfi srfi-88) in any
way, including via (@ ...) references will set 'postfix.

So stop doing that, and just accept #:foo style keywords on the Scheme
side.

Thanks to Zak for reporting the problem.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>