Add generic multi-kv assoc fallback
Add a fallback assoc specialization that handles multiple key value
pairs in terms of the single pair (assoc m k v) so that classes will
only have to define the latter when there's no worthwhile bulk
operation.

May not matter, but define the fallback with three required arguments
so that there's no chance of ambiguity with the two argument versions.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>