Migrate hash-map hash-set from pfds to fash
After spending quite a bit of time trying to fix what appear to be
bugs in either pfds hamts, or my understanding if it, Andy Wingo
pointed out the fash library.  Switch to that for now, since it
appears to work OK as-is.  Though we'll want to make some improvements
if we plan to keep it, i.e. dissoc support, which will also allow us
to go back to having a constant-time count.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>