Add initial, incomplete, support for *file*
Map *file* for now to (current-filename), and have it return nil when
the name isn't available.  This differs from the jvm where it returns
"NO_SOURCE_PATH".  I assume there's a justification for that, but I
haven't found it yet, and it does introduce the potential for
aliasing, so I thought we'd start with nil and just change it later if
that proves appropriate.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>