Add atom-meta and atom-alter-meta!
Rework (lokke scm atom) to support metadata.  Add atom-meta and
atom-alter-meta! to get and set it, and add some trivial tests.

Vaguely guess for now that atoms will have with increasing rarity
a validator, watchers, and metadata, and store the internal state in a
variable length vector reflecting that guess.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>