Improve metadata support (reader and runtime)
Add metadata support to hash-set, e.g. meta and with-meta.

Add reader metadata support for the [], {}, and #{} literals by
reworking the /lokke/reader-vector, /lokke/reader-hash-map,
/lokke/reader-hash-set, forms to require an initial metadata argument,
that can either be #nil or (recursively) a reader hash-set.  Adjust
the reader to place any pending metadata there.

Move some code from (lokke collection) to (lokke base collection) and
adjust various related dependencies in order to avoid cycles, given
the new requirements.

Add a new (lokke reader literal) module providing functions like
reader-hash-set? to better encapsulate operations on
the (/lokke/reader-* ...) forms, though we still don't fully
encapsulate those operations.  For example, we still explicitly match
syntaxes against the /lokke/reader-* symbols, etc.

Fix the handling of nil in some of the (lokke boot) syntaxes after
realizing that (exp ...) will match #nil, which was not particularly
surprising, but in the process, when substituted it becomes (), which
was surprising, i.e. explicitly test for #nil first, when necessary,
like this

  (syntax-case x ()
    ...
    ((_ nil x ...) (eq? #nil (syntax->datum nil)) something)
    ((_ (exp ...) x ...)) something-else)

Change synerr names from symbols to strings in (lokke boot) so that we
don't have to think about any interactions with the quote defined there.

Drop the meta and with-meta error fallbacks for unsupported types
since they weren't a useful improvement over the default behavior.

See the changes to README and DESIGN for additional information.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>