lokke: run guile directly (without a shell wrapper)
Run ./lokke directly via guile, eliminating the shell wrapper.  To do
so, automatically generate ./lokke to refer to the correct $(GUILE),
library, and module paths.  Look up the latter two relative to dirname
"$0", i.e. the source tree, but allow for the specification of literal
strings, which will support bin/lokke in a "make install" tree.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>