Support standalone installs via "make install"
Automatically generate the list of module files in Makefile.am via
git-ls-files.  Add a Makefile.am.in template and a gen-makefile
generator to handle the substituion.

Do this because automake currently has no easy way to omit or alter
the path of a file located in a subdirectory and we need to install
mod/x as LOKKE_EFFECTIVE_VERSION/x.  To make that change we have to
define a variable for each subdir in the tree and add each file to the
appropriate variable.

An alternate approach would be to switch to recursive automake, and
have a mod/Makefile.am.  I think that would probably allow us to
specify the name change once at the top-level (see guile's
module/Makefile.am for a likely example), but for now, let's stick
with a nonrecursive approach.

This also means that the build process requires git, and that we're
checking in one autogenerated file (Makefile.am).  In exchange, it
should be much more difficult to forget to update the file list.
Let's see how it goes.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>