Support (def :private ...)
Since we don't support symbol metadata, arrange
for (/lokke/reader-meta ...) forms that can't be appled to the
following data structure to be left alone by read-for-compiler so that
they can be matched by syntaxes.

Adjust def and defn to be able to match that reader metadata and
respect :private settings.  Add a /lokke/reader-meta syntax to the
boot environment that will warn about and drop any ignored reader
metadata for now, though I suspect we may eventually end up wanting to
drop the warning.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>