Replace cljl with lok; --#! with -0; execute !# preamble as scheme
For now, replace the cljl subcommand with run, and replace the cljl
convenience command with lok.  While lok and its equivalent "lokke
run" may still fill a niche similar to the JVM's clj command and share
options and behaviors when appropriate, they're not intended to match
clj as closely as cljl did.

Require a -- argument to separate *command-line-args* from interpreter
arguments, so that there are no (present or if new arguments are
added, future) ambiguities like this:

  echo '(prn "ok")' > -Stree
  clj -Stree

Start with --apply rather than -m since it's more flexible, though we
might add -m later.  clj's "-m x" can be expressed as "-a x/-main".

Allow arbitrary interleaving and repetition of load paths and
arguments like --apply, --load, and --eval, e.g.:

  lok -l x.clj -e "(a)" -a some/g -a other/h y.clj z.clj -- 1 2

Disallow load path names starting with a dash, i.e. any unknown name
should be specified via --load, e.g.

  lok --load "$1"

rather than

  lok "$1"

Drop the bespoke "lokke --#!" EDN #! ... !# preamble parsing in favor
of evaluating it as Scheme in a preamble-specific module defining
per-subcommand procedures (e.g. (run ...)), along with %0, %1, ... and
%&.  Make the preamble optional, and when it contains no expressions,
just execute the #! script in typical fashion.  See "man -l lokke.1"
for further details.

Replace --#! with -0 as the interpreter argument.  The intention is to
allow us to treat the number as a compatibility level (cf. debhelper)
which we can increment later if desired, and then be able to recognize
scripts expecting older behaviors, leaving open the possibility of
accommodating them.

As a hypothetical example, imagine we eventually wanted to change the
preamble handling in a backward-incompatible manner.  If we liked, we could
allow scripts to opt-in via "-1".

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>