Rework pr/print via pr-readable/pr-approchable; add *print-readably*
Base all printing on pr-readable and pr-approachable methods instead
of pr-on and print-on since clojure's pr method must really be both,
depending on the value of *print-readably*.  This arrangement also
prepares for the addition of a to-string method to support types whose
str values differ from their print values (e.g. #"x").

Prefer two separate basis methods so that types with differing output
don't have to switch on *print-readably* in every call, and choose
fairly unique names both so they're unlikely to conflict with future
upstream changes, and so it's easy to change them if we think of
something we prefer.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>