try: rewrite to dispatch all exceptions at the same level
Instead of nesting catch handlers that catch specific exception keys
to handle a try with multiple catch clauses, use one catch that
catches everything, and has a handler with a cond with one clause per
catch, that matches on the exception tag.

This should ensure that only the first matching clause catches, and it
prepares for support for a simple exception hierarchy, and for the
likely eventual move to implement all the exception handling in terms
of guile's newer raise-exception, with-exception-handler, and
exception objects.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>