Add trivial, experimental exception hierarchy
Add Error and arrange for the only supported "compatibility
exceptions", Throwable, Error, Exception, and ExceptionInfo, to
respect the jvm hierarchy, e.g. a catch for Exception will also
catch ExceptionInfo.

Map Error to all of the guile exception keys that map to &error in the
newer (guile 3+) exception handling infrastructure, so that, for
example, catching Error will also catch 'syntax-error.

This should be considered very experimental, and may well change
substantially, but for now, it allows some behavioral compatbility
with existing code.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>