Provide major-versioned command names, e.g. lokke-MAJOR
We may not stick with this, but for now, provided versioned flavors of
the installed commands, versioned at the semantic versioning "x"
level (for x.y.z).  If we really do respect semantic versioning, that
should be sufficient?

For now, install both command and command-MAJOR, but we might
eventually make the former a symlink to the latter, and of course
platforms may want to manage the unversioned name themselves,
e.g. Debian's update-alternatives.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>