Move syntax-quote from scheme to clojure
It's easier to implement on the clojure side, and it makes it easier
to match the expected semantics (which I haven't seen very formally
documented yet) closely enough, or to at least match the observed jvm
behavior.  It also allows us to simplify (lokke boot), e.g. it doesn't
need to be #:pure any more.

(And if it doesn't work out, we can always restore/enhance the scheme
 version...)

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>