Expose scoped syntax refs to the compiler
OK, this is a bit sneaky.  When testing some existing clojure macros,
I noticed that one of them returned something like this:

  `(... ~(list `fn [] ...))

which ends up producing a structure including the symbol
clojure.core/fn for the compiler, and the compiler has no idea that
symbol should actually be handled as a reference to the local fn
syntax binding.  As a result, the compiler (macroexpander) doesn't
expand it, and the resulting code crashes on an eventual attempt to
call fn as a function.

To fix that introuduce a new compiler literal

  (/lokke/ref-scoped-sym-if-syntax sym ns name)

and then transmogrify all scoped symbol references to that form when
shifting from the clojure domain to the compiler domain.  Switch them
back when going the other direction.

Define a compiler-specific /lokke/ref-scoped-sym-if-syntax macro that
if expanded (i.e. if it ends up in the evaluation position while
compiling) will look up the original symbol in the relevant module,
and if bound to a macro, expand into an (@ ...) reference to the
target indicated by the scoped symbol.  That allows the macro to then
expand the form (via (@ (lokke ns clojure core) fn) in the example) as
expected.  Otherwise, if the scoped symbol doesn't refer to a macro
binding, then have /lokke/ref-scoped-sym-if-syntax just expand to the
original scoped symbol.

Teach clj-quote how to handle /lokke/ref-scoped-sym-if-syntax as well,
i.e. make any attempt to quote that form expand to the original scoped
symbol.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>