Move quote to (lokke base quote); stop using it in (lokke core)
Move the clojure quote to (lokke base quote) and stop using the
clojure quote in (lokke core).  i.e. export it as quote, but let ' and
quote in core.scm be scheme quote, since we didn't need clojure quote,
and there's less potential for accidental complexity/confusion.

This also supports the forthcoming migration of syntax-quote from
scheme to clojure.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>