Make (seq map) eager for now to match (keys map) order
Clojure requires that a map's seq and keys ordering be the same, so
for now, change seq to work like keys.  This means it's no longer
eager, though it may be a lot more efficient overall than the previous
"incremental dissoc" approach (which incrementally built an entirely
new map with "not found" entries for every key given the current fash
limitations, i.e. no dissoc).

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>