Don't accidentally, implicitly :refer :all
Oops.  The code that generates the guile module #:select for a given
ns form acted as if every require included a :refer :all because it
defaulted the select value to #f all the time (for both use and
require), when it should have been choosing #f for use and () for
require, since guile treats a "#:select #f" like :refer :all.

i.e. in guile (use-modules (srfi srfi-1)) imports all of the srfi-1
exports into the current module, and (use-modules ((srfi srfi-1)
\#:select #f) behaves exactly the same way, while (use-modules ((srfi
srfi-1) #:select ()) imports nothing.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>